Move board to an external component
