+++ src/components/index.js
@@ -1,0 +1,1 @@
+export { default as Board } from "./Board/Board";
